+++ zed_components/src/zed_camera/src/zed_camera_component.cpp
@@ -139,9 +139,9 @@ ZedCamera::~ZedCamera()
     stopObjDetect();
   }
 
-  if (mBodyTrackRunning) {
-    std::lock_guard<std::mutex> lock(mBodyTrackMutex);
-    stopBodyTracking();
+  if (mBodyTrkRunning) {
+    std::lock_guard<std::mutex> lock(mBodyTrkMutex);
+    stopBodyTrking();
   }
 
   if (mSpatialMappingRunning) {
@@ -407,11 +407,11 @@ void ZedCamera::initParameters()
   if (!mDepthDisabled) {
     if (sl_tools::isObjDetAvailable(mCamUserModel)) {
       getOdParams();
-      getBodyTrackParams();
+      getBodyTrkParams();
     }
   } else {
     mObjDetEnabled = false;
-    mBodyTrackEnabled = false;
+    mBodyTrkEnabled = false;
   }
 }
 
@@ -514,10 +514,10 @@ void ZedCamera::getDebugParams()
     get_logger(), " * Debug Object Detection: %s",
     mDebugObjectDet ? "TRUE" : "FALSE");
 
-  getParam("debug.debug_body_tracking", mDebugBodyTrack, mDebugBodyTrack);
+  getParam("debug.debug_body_tracking", mDebugBodyTrk, mDebugBodyTrk);
   RCLCPP_INFO(
     get_logger(), " * Debug Body Tracking: %s",
-    mDebugBodyTrack ? "TRUE" : "FALSE");
+    mDebugBodyTrk ? "TRUE" : "FALSE");
 
 #ifdef WITH_TM
   getParam("debug.debug_terrain_mapping", mDebugTerrainMapping, mDebugTerrainMapping);
@@ -528,7 +528,7 @@ void ZedCamera::getDebugParams()
 
   mDebugMode = mDebugCommon || mDebugVideoDepth || mDebugPointCloud ||
     mDebugPosTracking || mDebugGnss || mDebugSensors || mDebugMapping ||
-    mDebugObjectDet || mDebugBodyTrack;
+    mDebugObjectDet || mDebugBodyTrk;
 
 #ifdef WITH_TM
   mDebugMode |= mDebugTerrainMapping;
@@ -1621,7 +1621,7 @@ void ZedCamera::getOdParams()
     get_logger(), " * Obj. Det. QoS Durability: %s", sl_tools::qos2str(qos_durability).c_str());
 }
 
-void ZedCamera::getBodyTrackParams()
+void ZedCamera::getBodyTrkParams()
 {
   rclcpp::Parameter paramVal;
   std::string paramName;
@@ -1640,9 +1640,9 @@ void ZedCamera::getBodyTrackParams()
     return;
   }
 
-  getParam("body_tracking.bt_enabled", mBodyTrackEnabled, mBodyTrackEnabled);
+  getParam("body_tracking.bt_enabled", mBodyTrkEnabled, mBodyTrkEnabled);
   RCLCPP_INFO_STREAM(
-    get_logger(), " * Body Track. enabled: " << (mBodyTrackEnabled ? "TRUE" : "FALSE"));
+    get_logger(), " * Body Track. enabled: " << (mBodyTrkEnabled ? "TRUE" : "FALSE"));
 
   bool matched = false;
 
@@ -1657,7 +1657,7 @@ void ZedCamera::getBodyTrackParams()
     std::replace(test_model_str.begin(), test_model_str.end(), ' ', '_'); // Replace spaces with underscores to match the YAML setting
     //DEBUG_BT(" Comparing '%s' to '%s'", test_model_str.c_str(), model_str.c_str());
     if (model_str == test_model_str) {
-      mBodyTrackModel = test_model;
+      mBodyTrkModel = test_model;
       matched = true;
       break;
     }
@@ -1670,7 +1670,7 @@ void ZedCamera::getBodyTrackParams()
   }
   RCLCPP_INFO_STREAM(
     get_logger(),
-    " * Body Track. model: " << sl::toString(mBodyTrackModel).c_str());
+    " * Body Track. model: " << sl::toString(mBodyTrkModel).c_str());
 
   std::string fmt_str = "BODY_70";
   getParam("body_tracking.body_format", fmt_str, fmt_str);
@@ -1683,7 +1683,7 @@ void ZedCamera::getBodyTrackParams()
     std::replace(test_fmt_str.begin(), test_fmt_str.end(), ' ', '_'); // Replace spaces with underscores to match the YAML setting
     //DEBUG_BT(" Comparing '%s' to '%s'", test_fmt_str.c_str(), test_fmt.c_str());
     if (fmt_str == test_fmt_str) {
-      mBodyTrackFmt = test_fmt;
+      mBodyTrkFmt = test_fmt;
       matched = true;
       break;
     }
@@ -1696,16 +1696,16 @@ void ZedCamera::getBodyTrackParams()
   }
   RCLCPP_INFO_STREAM(
     get_logger(),
-    " * Body Track. format: " << sl::toString(mBodyTrackFmt).c_str());
+    " * Body Track. format: " << sl::toString(mBodyTrkFmt).c_str());
 
   getParam(
-    "body_tracking.allow_reduced_precision_inference", mBodyTrackReducedPrecision,
-    mBodyTrackReducedPrecision);
+    "body_tracking.allow_reduced_precision_inference", mBodyTrkReducedPrecision,
+    mBodyTrkReducedPrecision);
   RCLCPP_INFO_STREAM(
     get_logger(),
     " * Body Track. allow reduced precision: " << (mObjDetReducedPrecision ? "TRUE" : "FALSE"));
   getParam(
-    "body_tracking.max_range", mBodyTrackMaxRange, mBodyTrackMaxRange,
+    "body_tracking.max_range", mBodyTrkMaxRange, mBodyTrkMaxRange,
     " * Body Track. maximum range [m]: ");
 
   std::string body_sel_str = "FULL";
@@ -1723,7 +1723,7 @@ void ZedCamera::getBodyTrackParams()
     std::replace(test_body_sel_str.begin(), test_body_sel_str.end(), ' ', '_'); // Replace spaces with underscores to match the YAML setting
     DEBUG_BT(" Comparing '%s' to '%s'", test_body_sel_str.c_str(), body_sel_str.c_str());
     if (body_sel_str == test_body_sel_str) {
-      mBodyTrackKpSelection = test_kp_sel;
+      mBodyTrkKpSelection = test_kp_sel;
       matched = true;
       break;
     }
@@ -1735,24 +1735,24 @@ void ZedCamera::getBodyTrackParams()
         "'. Using the default value.");
   }
 
-  getParam("body_tracking.enable_body_fitting", mBodyTrackFitting, mBodyTrackFitting);
+  getParam("body_tracking.enable_body_fitting", mBodyTrkFitting, mBodyTrkFitting);
   RCLCPP_INFO_STREAM(
-    get_logger(), " * Body fitting: " << (mBodyTrackFitting ? "TRUE" : "FALSE"));
+    get_logger(), " * Body fitting: " << (mBodyTrkFitting ? "TRUE" : "FALSE"));
 
-  getParam("body_tracking.enable_tracking", mBodyTrackEnableTracking, mBodyTrackEnableTracking);
+  getParam("body_tracking.enable_tracking", mBodyTrkEnableTracking, mBodyTrkEnableTracking);
   RCLCPP_INFO_STREAM(
-    get_logger(), " * Body joints tracking: " << (mBodyTrackEnableTracking ? "TRUE" : "FALSE"));
+    get_logger(), " * Body joints tracking: " << (mBodyTrkEnableTracking ? "TRUE" : "FALSE"));
 
   getParam(
-    "body_tracking.prediction_timeout_s", mBodyTrackPredTimeout, mBodyTrackPredTimeout,
+    "body_tracking.prediction_timeout_s", mBodyTrkPredTimeout, mBodyTrkPredTimeout,
     " * Body Track. prediction timeout [sec]: ");
 
   getParam(
-    "body_tracking.confidence_threshold", mBodyTrackConfThresh, mBodyTrackConfThresh,
+    "body_tracking.confidence_threshold", mBodyTrkConfThresh, mBodyTrkConfThresh,
     " * Body Track. confidence thresh.: ", true);
 
   getParam(
-    "body_tracking.minimum_keypoints_threshold", mBodyTrackMinKp, mBodyTrackMinKp,
+    "body_tracking.minimum_keypoints_threshold", mBodyTrkMinKp, mBodyTrkMinKp,
     " * Body Track. min. KP thresh.: ", true);
 
   // ------------------------------------------
@@ -2389,7 +2389,7 @@ rcl_interfaces::msg::SetParametersResult ZedCamera::callback_paramChange(
         break;
       }
 
-      mBodyTrackConfThresh = val;
+      mBodyTrkConfThresh = val;
 
       RCLCPP_INFO_STREAM(
         get_logger(), "Parameter '" << param.get_name() << "' correctly set to " << val);
@@ -2412,7 +2412,7 @@ rcl_interfaces::msg::SetParametersResult ZedCamera::callback_paramChange(
         break;
       }
 
-      mBodyTrackMinKp = val;
+      mBodyTrkMinKp = val;
 
       RCLCPP_INFO_STREAM(
         get_logger(), "Parameter '" << param.get_name() << "' correctly set to " << val);
@@ -2699,7 +2699,7 @@ void ZedCamera::initPublishers()
   mObjectDetTopic = mTopicRoot + object_det_topic_root + "/objects";
 
   std::string body_trk_topic_root = "body_trk";
-  mBodyTrackTopic = mTopicRoot + body_trk_topic_root + "/skeletons";
+  mBodyTrkTopic = mTopicRoot + body_trk_topic_root + "/skeletons";
 
   std::string confImgRoot = "confidence";
   std::string conf_map_topic_name = "confidence_map";
@@ -3311,8 +3311,8 @@ bool ZedCamera::startCamera()
   mPcProcMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mObjDetPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mObjDetElabMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
-  mBodyTrackPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
-  mBodyTrackElabMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
+  mBodyTrkPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
+  mBodyTrkElabMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mImuPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mSensPubRate);
   mBaroPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mSensPubRate);
   mMagPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mSensPubRate);
@@ -3493,7 +3493,7 @@ bool ZedCamera::startPosTracking()
     RCLCPP_WARN(
       get_logger(),
       "Cannot start Positional Tracking if "
-      "`depth.quality` is set to `0` [NONE]");
+      "`depth.depth_mode` is set to `0` [NONE]");
     return false;
   }
 
@@ -3608,7 +3608,7 @@ bool ZedCamera::start3dMapping()
 {
   DEBUG_MAP("start3dMapping");
   if (mDepthDisabled) {
-    RCLCPP_WARN(get_logger(), "Cannot start 3D Mapping if `depth.quality` is set to `0` [NONE]");
+    RCLCPP_WARN(get_logger(), "Cannot start 3D Mapping if `depth.depth_mode` is set to `0` [NONE]");
     return false;
   }
 
@@ -3732,7 +3732,7 @@ bool ZedCamera::startTerrainMapping()
   if (mDepthDisabled) {
     RCLCPP_WARN(
       get_logger(),
-      "Cannot start Terrain Mapping if `depth.quality` is set to `0` [NONE]");
+      "Cannot start Terrain Mapping if `depth.depth_mode` is set to `0` [NONE]");
     return false;
   }
   if (!mTerrainMappingEnabled) {
@@ -3795,7 +3795,7 @@ bool ZedCamera::startObjDetect()
     RCLCPP_WARN(
       get_logger(),
       "Cannot start Object Detection if "
-      "`depth.quality` is set to `0` [NONE]");
+      "`depth.depth_mode` is set to `0` [NONE]");
     return false;
   }
 
@@ -3860,13 +3860,6 @@ bool ZedCamera::startObjDetect()
     RCLCPP_INFO_STREAM(get_logger(), "Advertised on topic " << mPubObjDet->get_topic_name());
   }
 
-  if (!mPubBodyTrack) {
-    mPubBodyTrack = create_publisher<zed_interfaces::msg::ObjectsStamped>(
-      mBodyTrackTopic,
-      mBodyTrkQos);
-    RCLCPP_INFO_STREAM(get_logger(), "Advertised on topic " << mPubBodyTrack->get_topic_name());
-  }
-
   mObjDetRunning = true;
   return true;
 }
@@ -3895,16 +3888,99 @@ void ZedCamera::stopObjDetect()
   }
 }
 
-bool ZedCamera::startBodyTracking()
+bool ZedCamera::startBodyTrking()
 {
-  DEBUG_BT("startBodyTracking");
+  DEBUG_BT("startBodyTrking");
 
-  return false;
+  if (!sl_tools::isObjDetAvailable(mCamRealModel)) {
+    RCLCPP_ERROR(
+      get_logger(),
+      "Body Tracking not started. The camera model does not support it with the current version "
+      "of the SDK");
+    return false;
+  }
+
+  DEBUG_BT("Body Tracking available");
+
+  if (mDepthDisabled) {
+    RCLCPP_WARN(
+      get_logger(),
+      "Cannot start Body Tracking if "
+      "`depth.depth_mode` is set to `0` [NONE]");
+    return false;
+  }
+
+  if (!mBodyTrkEnabled) {
+    DEBUG_BT("Body Tracking not enabled -> NOT STARTING");
+    return false;
+  }
+
+  if (!mCamera2BaseTransfValid || !mSensor2CameraTransfValid || !mSensor2BaseTransfValid) {
+    DEBUG_OD("Tracking transforms not yet ready, Body Tracking starting postponed");
+    return false;
+  }
+
+  RCLCPP_INFO(get_logger(), "*** Starting Body Tracking ***");
+
+  sl::BodyTrackingParameters bt_p;
+  bt_p.allow_reduced_precision_inference = mBodyTrkReducedPrecision;
+  bt_p.body_format = mBodyTrkFmt;
+  bt_p.body_selection = mBodyTrkKpSelection;
+  bt_p.detection_model = mBodyTrkModel;
+  bt_p.enable_body_fitting = mBodyTrkFitting;
+  bt_p.enable_segmentation = false;
+  bt_p.enable_tracking = mBodyTrkEnableTracking;
+  bt_p.image_sync = true;
+  bt_p.max_range = mBodyTrkMaxRange;
+  bt_p.prediction_timeout_s = mBodyTrkPredTimeout;
+
+  // TODO(Walter) Consider adding future support for `instance_module_id`
+  //bt_p.instance_module_id = 0;
+
+  sl::ERROR_CODE btError = mZed.enableBodyTracking(bt_p);
+
+  if (btError != sl::ERROR_CODE::SUCCESS) {
+    RCLCPP_ERROR_STREAM(get_logger(), "Body Tracking error: " << sl::toString(btError));
+
+    mBodyTrkRunning = false;
+    return false;
+  }
+
+  DEBUG_BT("Body Tracking enabled");
+
+  if (!mPubBodyTrk) {
+    mPubBodyTrk = create_publisher<zed_interfaces::msg::ObjectsStamped>(mBodyTrkTopic, mBodyTrkQos);
+    RCLCPP_INFO_STREAM(get_logger(), "Advertised on topic " << mPubBodyTrk->get_topic_name());
+  }
+
+  DEBUG_BT("Body Tracking publisher created");
+
+  mBodyTrkRunning = true;
+  return true;
 }
 
-void ZedCamera::stopBodyTracking()
+void ZedCamera::stopBodyTrking()
 {
+  if (mBodyTrkRunning) {
+    RCLCPP_INFO(get_logger(), "*** Stopping Body Tracking ***");
+    mBodyTrkRunning = false;
+    mBodyTrkEnabled = false;
+    mZed.disableBodyTracking();
 
+    // ----> Send an empty message to indicate that no more objects are tracked
+    // (e.g clean Rviz2)
+    objDetMsgPtr objMsg = std::make_unique<zed_interfaces::msg::ObjectsStamped>();
+
+    objMsg->header.stamp = mFrameTimestamp;
+    objMsg->header.frame_id = mLeftCamFrameId;
+
+    objMsg->objects.clear();
+
+    DEBUG_STREAM_OD("Publishing EMPTY OBJ message " << mPubBodyTrk->get_topic_name());
+    mPubBodyTrk->publish(std::move(objMsg));
+    // <---- Send an empty message to indicate that no more objects are tracked
+    // (e.g clean Rviz2)
+  }
 }
 
 bool ZedCamera::startSvoRecording(std::string & errMsg)
@@ -4414,14 +4490,14 @@ void ZedCamera::threadFunc_zedGrab()
 
     // ----> Check for Body Tracking requirement
     if (!mDepthDisabled) {
-      mBodyTrackMutex.lock();
-      if (mBodyTrackEnabled && !mBodyTrackRunning) {
-        startBodyTracking();
+      mBodyTrkMutex.lock();
+      if (mBodyTrkEnabled && !mBodyTrkRunning) {
+        startBodyTrking();
         if (!sl_tools::isObjDetAvailable(mCamRealModel)) {
-          mBodyTrackEnabled = false;
+          mBodyTrkEnabled = false;
         }
       }
-      mBodyTrackMutex.unlock();
+      mBodyTrkMutex.unlock();
     }
     // ----> Check for Object Detection requirement
 
@@ -4608,11 +4684,11 @@ void ZedCamera::threadFunc_zedGrab()
     }
 
     if (!mDepthDisabled) {
-      mBodyTrackMutex.lock();
-      if (mBodyTrackRunning) {
-        processSkeletons(mFrameTimestamp);
+      mBodyTrkMutex.lock();
+      if (mBodyTrkRunning) {
+        processBodies(mFrameTimestamp);
       }
-      mBodyTrackMutex.unlock();
+      mBodyTrkMutex.unlock();
     }
 
     // Diagnostic statistics update
@@ -6268,36 +6344,20 @@ void ZedCamera::processDetectedObjects(rclcpp::Time t)
 
     memcpy(&(objMsg->objects[idx].dimensions_3d[0]), &(data.dimensions[0]), 3 * sizeof(float));
 
-    // TODO(Walter) for SDK v4 integration the OD module must be reworked!!!
+    if (data.head_bounding_box_2d.size() == 4) {
+      memcpy(
+        &(objMsg->objects[idx].head_bounding_box_2d.corners[0]), &(data.head_bounding_box_2d[0]),
+        8 * sizeof(unsigned int));
+    }
+    if (data.head_bounding_box.size() == 8) {
+      memcpy(
+        &(objMsg->objects[idx].head_bounding_box_3d.corners[0]), &(data.head_bounding_box[0]),
+        24 * sizeof(float));
+    }
 
-//     objMsg->objects[idx].body_format = static_cast<uint8_t>(mObjDetBodyFmt);
+    memcpy(&(objMsg->objects[idx].head_position[0]), &(data.head_position[0]), 3 * sizeof(float));
 
-//       if (data.head_bounding_box_2d.size() == 4) {
-//         memcpy(
-//           &(objMsg->objects[idx].head_bounding_box_2d.corners[0]), &(data.head_bounding_box_2d[0]),
-//           8 * sizeof(unsigned int));
-//       }
-//       if (data.head_bounding_box.size() == 8) {
-//         memcpy(
-//           &(objMsg->objects[idx].head_bounding_box_3d.corners[0]), &(data.head_bounding_box[0]),
-//           24 * sizeof(float));
-//       }
-//       memcpy(&(objMsg->objects[idx].head_position[0]), &(data.head_position[0]), 3 * sizeof(float));
-
-//       uint8_t kp_size = data.keypoint_2d.size();
-//       if (kp_size == 18 || kp_size == 34) {
-//         memcpy(
-//           &(objMsg->objects[idx].skeleton_2d.keypoints[0]), &(data.keypoint_2d[0]),
-//           2 * kp_size * sizeof(float));
-
-//         memcpy(
-//           &(objMsg->objects[idx].skeleton_3d.keypoints[0]), &(data.keypoint[0]),
-//           3 * kp_size * sizeof(float));
-//       }
-//     } else {
-//       objMsg->objects[idx].skeleton_available = false;
-//     }
-
+    objMsg->objects[idx].skeleton_available = false;
     // at the end of the loop
     idx++;
   }
@@ -6312,34 +6372,146 @@ void ZedCamera::processDetectedObjects(rclcpp::Time t)
   // <---- Diagnostic information update
 }
 
-void ZedCamera::processSkeletons(rclcpp::Time t)
+void ZedCamera::processBodies(rclcpp::Time t)
 {
-  DEBUG_BT("processSkeletons");
+  //DEBUG_BT("processBodies");
+
   size_t bt_sub_count = 0;
 
   try {
-    bt_sub_count = count_subscribers(mPubBodyTrack->get_topic_name());
+    bt_sub_count = count_subscribers(mPubBodyTrk->get_topic_name());
   } catch (...) {
     rcutils_reset_error();
-    DEBUG_STREAM_OD("processSkeletons: Exception while counting subscribers");
+    DEBUG_STREAM_OD("processBodies: Exception while counting subscribers");
     return;
   }
 
   if (bt_sub_count < 1) {
-    mObjDetSubscribed = false;
+    mBodyTrkSubscribed = false;
     return;
   }
 
   sl_tools::StopWatch btElabTimer(get_clock());
   static sl_tools::StopWatch btFreqTimer(get_clock());
 
+  mBodyTrkSubscribed = true;
+
+  // ----> Process realtime dynamic parameters
   sl::BodyTrackingRuntimeParameters bt_params_rt;
-  bt_params_rt.detection_confidence_threshold;
-  bt_params_rt.minimum_keypoints_threshold;
+  bt_params_rt.detection_confidence_threshold = mBodyTrkConfThresh;
+  bt_params_rt.minimum_keypoints_threshold = mBodyTrkMinKp;
+  // <---- Process realtime dynamic parameters
+
+  sl::Bodies bodies;
+  sl::ERROR_CODE btRes = mZed.retrieveBodies(bodies, bt_params_rt);
+
+  if (btRes != sl::ERROR_CODE::SUCCESS) {
+    RCLCPP_WARN_STREAM(get_logger(), "Body Tracking error: " << sl::toString(btRes));
+    return;
+  }
+
+  if (!bodies.is_new) {  // Async body tracking. Update data only if new
+    // detection is available
+    DEBUG_BT("No new bodies detected");
+    return;
+  }
+
+  size_t bodyCount = bodies.body_list.size();
+
+  DEBUG_STREAM_BT("Detected " << bodyCount << " bodies");
+
+  objDetMsgPtr bodyMsg = std::make_unique<zed_interfaces::msg::ObjectsStamped>();
+
+  bodyMsg->header.stamp = t;
+  bodyMsg->header.frame_id = mLeftCamFrameId;
+
+  bodyMsg->objects.resize(bodyCount);
+
+  size_t idx = 0;
+  for (auto body : bodies.body_list) {
+    std::string label = "Body_";
+    label += std::to_string(body.id);
+    DEBUG_STREAM_BT("Processing body: " << label);
+    bodyMsg->objects[idx].label = sl::String(label.c_str());
+    bodyMsg->objects[idx].sublabel = "";
+    bodyMsg->objects[idx].label_id = body.id;
+    bodyMsg->objects[idx].confidence = body.confidence;
+    DEBUG_BT(" - Info OK");
+
+    memcpy(&(bodyMsg->objects[idx].position[0]), &(body.position[0]), 3 * sizeof(float));
+    memcpy(
+      &(bodyMsg->objects[idx].position_covariance[0]), &(body.position_covariance[0]),
+      6 * sizeof(float));
+    memcpy(&(bodyMsg->objects[idx].velocity[0]), &(body.velocity[0]), 3 * sizeof(float));
+    DEBUG_BT(" - Pos/Cov/Speed OK");
+
+    bodyMsg->objects[idx].tracking_available = mBodyTrkEnableTracking;
+    bodyMsg->objects[idx].tracking_state = static_cast<int8_t>(body.tracking_state);
+    bodyMsg->objects[idx].action_state = static_cast<int8_t>(body.action_state);
+    DEBUG_BT(" - Status OK");
+
+    if (body.bounding_box_2d.size() == 4) {
+      memcpy(
+        &(bodyMsg->objects[idx].bounding_box_2d.corners[0]), &(body.bounding_box_2d[0]),
+        8 * sizeof(unsigned int));
+    }
+    DEBUG_BT(" - BBox 2D OK");
+    if (body.bounding_box.size() == 8) {
+      memcpy(
+        &(bodyMsg->objects[idx].bounding_box_3d.corners[0]), &(body.bounding_box[0]),
+        24 * sizeof(float));
+    }
+    DEBUG_BT(" - BBox 3D OK");
+
+    memcpy(&(bodyMsg->objects[idx].dimensions_3d[0]), &(body.dimensions[0]), 3 * sizeof(float));
+    DEBUG_BT(" - Dims OK");
+
+    bodyMsg->objects[idx].body_format = static_cast<uint8_t>(mBodyTrkFmt);
+
+    if (body.head_bounding_box_2d.size() == 4) {
+      memcpy(
+        &(bodyMsg->objects[idx].head_bounding_box_2d.corners[0]), &(body.head_bounding_box_2d[0]),
+        8 * sizeof(unsigned int));
+    }
+    if (body.head_bounding_box.size() == 8) {
+      memcpy(
+        &(bodyMsg->objects[idx].head_bounding_box_3d.corners[0]), &(body.head_bounding_box[0]),
+        24 * sizeof(float));
+    }
+
+    memcpy(&(bodyMsg->objects[idx].head_position[0]), &(body.head_position[0]), 3 * sizeof(float));
+
+    bodyMsg->objects[idx].skeleton_available = true;
+
+    uint8_t kp_size = body.keypoint_2d.size();
+    if (kp_size == 18 || kp_size == 34) {
+      memcpy(
+        &(bodyMsg->objects[idx].skeleton_2d.keypoints[0]), &(body.keypoint_2d[0]),
+        2 * kp_size * sizeof(float));
+
+      memcpy(
+        &(bodyMsg->objects[idx].skeleton_3d.keypoints[0]), &(body.keypoint[0]),
+        3 * kp_size * sizeof(float));
+    }
+
+
+    // ----------------------------------
+    // at the end of the loop
+
+    // TODO(Walter) Add support for
+    //body.global_root_orientation;
+    //body.local_orientation_per_joint;
+    //body.local_orientation_per_joint;
+
+    idx++;
+  }
+
+  DEBUG_STREAM_OD("Publishing BODY TRK message");
+  mPubBodyTrk->publish(std::move(bodyMsg));
 
   // ----> Diagnostic information update
-  mBodyTrackElabMean_sec->addValue(btElabTimer.toc());
-  mBodyTrackPeriodMean_sec->addValue(btFreqTimer.toc());
+  mBodyTrkElabMean_sec->addValue(btElabTimer.toc());
+  mBodyTrkPeriodMean_sec->addValue(btFreqTimer.toc());
   btFreqTimer.tic();
   // <---- Diagnostic information update
 }
@@ -7642,14 +7814,14 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
         stat.add("Object Detection", "INACTIVE");
       }
 
-      if (mBodyTrackRunning) {
-        if (mBodyTrackSubscribed) {
-          double freq = 1. / mBodyTrackPeriodMean_sec->getAvg();
+      if (mBodyTrkRunning) {
+        if (mBodyTrkSubscribed) {
+          double freq = 1. / mBodyTrkPeriodMean_sec->getAvg();
           double freq_perc = 100. * freq / mPubFrameRate;
           stat.addf("Body Tracking", "Mean Frequency: %.3f Hz  (%.1f%%)", freq, freq_perc);
           stat.addf(
             "Body Tracking", "Processing Time: %.3f sec (Max. %.3f sec)",
-            mBodyTrackElabMean_sec->getAvg(), 1. / mCamGrabFrameRate);
+            mBodyTrkElabMean_sec->getAvg(), 1. / mCamGrabFrameRate);
         } else {
           stat.add("Body Tracking", "Active, topic not subscribed");
         }